Partial separation of files
