+++ src/oauth.github.js
@@ -1,0 +1,61 @@
+angular.module('oauth.github', ['oauth.utils'])
+  .factory('$github', github);
+
+function github($q, $http, $cordovaUtility) {
+  return { signin: oauthGithub }
+
+  /*
+   * Sign into the GitHub service
+   *
+   * @param    string clientId
+   * @param    string clientSecret
+   * @param    array appScope
+   * @param    object options
+   * @return   promise
+   */
+  function oauthGithub(clientId, clientSecret, appScope, options) {
+    var deferred = $q.defer();
+    if(window.cordova) {
+      var cordovaMetadata = cordova.require("cordova/plugin_list").metadata;
+      if($cordovaOauthUtility.isInAppBrowserInstalled(cordovaMetadata) === true) {
+        var redirect_uri = "http://localhost/callback";
+        if(options !== undefined) {
+          if(options.hasOwnProperty("redirect_uri")) {
+            redirect_uri = options.redirect_uri;
+          }
+        }
+        var browserRef = window.open('https://github.com/login/oauth/authorize?client_id=' + clientId + '&redirect_uri=' + redirect_uri + '&scope=' + appScope.join(","), '_blank', 'location=no,clearsessioncache=yes,clearcache=yes');
+        browserRef.addEventListener('loadstart', function(event) {
+          if((event.url).indexOf(redirect_uri) === 0) {
+            requestToken = (event.url).split("code=")[1];
+            $http.defaults.headers.post['Content-Type'] = 'application/x-www-form-urlencoded';
+            $http.defaults.headers.post.accept = 'application/json';
+            $http({method: "post", url: "https://github.com/login/oauth/access_token", data: "client_id=" + clientId + "&client_secret=" + clientSecret + "&redirect_uri=" + redirect_uri + "&code=" + requestToken })
+              .success(function(data) {
+                deferred.resolve(data);
+              })
+              .error(function(data, status) {
+                deferred.reject("Problem authenticating");
+              })
+              .finally(function() {
+                setTimeout(function() {
+                    browserRef.close();
+                }, 10);
+              });
+          }
+        });
+        browserRef.addEventListener('exit', function(event) {
+            deferred.reject("The sign in flow was canceled");
+        });
+      } else {
+          deferred.reject("Could not find InAppBrowser plugin");
+      }
+    } else {
+      deferred.reject("Cannot authenticate via a web browser");
+    }
+
+    return deferred.promise;
+  }
+}
+
+$github.$inject = ['$q', '$http', '$cordovaUtility'];
