+++ src/oauth.linkedin.js
@@ -1,0 +1,61 @@
+angular.module('oauth.linkedin', ['oauth.utils'])
+  .factory('$linkedin', linkedin);
+
+function linkedin($q, $http, $cordovaUtility) {
+  return { signin: oauthLinkedin }
+
+  /*
+   * Sign into the LinkedIn service
+   *
+   * @param    string clientId
+   * @param    string clientSecret
+   * @param    array appScope
+   * @param    string state
+   * @param    object options
+   * @return   promise
+   */
+  function oauthLinkedin(clientId, clientSecret, appScope, state, options) {
+    var deferred = $q.defer();
+    if(window.cordova) {
+      var cordovaMetadata = cordova.require("cordova/plugin_list").metadata;
+      if($cordovaOauthUtility.isInAppBrowserInstalled(cordovaMetadata) === true) {
+        var redirect_uri = "http://localhost/callback";
+        if(options !== undefined) {
+          if(options.hasOwnProperty("redirect_uri")) {
+            redirect_uri = options.redirect_uri;
+          }
+        }
+
+        var browserRef = window.open('https://www.linkedin.com/uas/oauth2/authorization?client_id=' + clientId + '&redirect_uri=' + redirect_uri + '&scope=' + appScope.join(" ") + '&response_type=code&state=' + state, '_blank', 'location=no,clearsessioncache=yes,clearcache=yes');
+        browserRef.addEventListener('loadstart', function(event) {
+          if((event.url).indexOf(redirect_uri) === 0) {
+            requestToken = (event.url).split("code=")[1].split("&")[0];
+            $http.defaults.headers.post['Content-Type'] = 'application/x-www-form-urlencoded';
+            $http({method: "post", url: "https://www.linkedin.com/uas/oauth2/accessToken", data: "client_id=" + clientId + "&client_secret=" + clientSecret + "&redirect_uri=" + redirect_uri + "&grant_type=authorization_code" + "&code=" + requestToken })
+              .success(function(data) {
+                deferred.resolve(data);
+              })
+              .error(function(data, status) {
+                deferred.reject("Problem authenticating");
+              })
+              .finally(function() {
+                setTimeout(function() {
+                    browserRef.close();
+                }, 10);
+              });
+          }
+        });
+        browserRef.addEventListener('exit', function(event) {
+          deferred.reject("The sign in flow was canceled");
+        });
+      } else {
+        deferred.reject("Could not find InAppBrowser plugin");
+      }
+    } else {
+      deferred.reject("Cannot authenticate via a web browser");
+    }
+    return deferred.promise;
+  }
+}
+
+$linkedin.$inject = ['$q', '$http', '$cordovaUtility'];
